requirements.txt
# FastAPI PDF Compliance Checker - Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# PDF Processing
PyMuPDF==1.23.8

# Machine Learning / AI
sentence-transformers==2.2.2
rank-bm25==0.2.2
numpy==1.24.3
google-generativeai==0.3.2
openai==2.8.0

# Vector Database
qdrant-client==1.7.0

# HTTP Client (for testing)
requests==2.31.0

# Optional: Development Tools
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
